fin crud client+ rajout email customerController

+++ front/src/components/App.js
@@ -37,7 +37,7 @@ const App = () => {
   const [viewComponent, setViewComponent] = useState("");
 
   const [isUpdateCustomer, setIsUpdateCustomer] = useState(null);
-  const [selectedCustomer, setSelectedCustomer]= useState([]);
+  const [selectedCustomer, setSelectedCustomer] = useState([]);
 
 
   class handleCarNCoContext {
@@ -254,13 +254,13 @@ const App = () => {
       setShowFormClient(false);
     };
 
-    static handleClickOpenNewClientByCommande = (byCommande,isUpdateCustomer,idCustomer =null) => {
+    static handleClickOpenNewClientByCommande = (byCommande, isUpdateCustomer, idCustomer = null) => {
       setIsUpdateCustomer(isUpdateCustomer)
       setShowFormClient(true);
       setShowFormClientByCommande(byCommande);
-      if(isUpdateCustomer != null){
+      if (isUpdateCustomer != null) {
         const customers_copy = [...listCustomers];
-        const formDataCustomer = customers_copy.filter(customer =>{return idCustomer == customer._id});
+        const formDataCustomer = customers_copy.filter(customer => { return idCustomer == customer._id });
         setSelectedCustomer(formDataCustomer);
       }
 
@@ -273,6 +273,7 @@ const App = () => {
         firstName: e.target.clientFirstName.value,
         email: e.target.clientEmail.value,
         address: {
+          street: e.target.clientStreet.value,
           zipCode: e.target.clientZipCode.value,
           city: e.target.clientCity.value,
         },
@@ -286,20 +287,11 @@ const App = () => {
             this.handleClickNextClientByCommande(result);
           }
           this.getAllCustomers();
-          this.openToast(
-            "success",
-            "3000",
-            "Customer bien ajouté"
-          );
+          this.openToast("success","3000","Customer bien ajouté");
           setTimeout(this.hideLoader, 1000);
         })
         .catch((error) => {
-          this.openToast(
-            "danger",
-            "3000",
-            error.response.data.error,
-            error.response.data.code
-          );
+          this.openToast( "danger","3000",error.response.data.error,error.response.data.code);
           setTimeout(this.hideLoader, 1000);
         });
 
@@ -311,20 +303,19 @@ const App = () => {
     }
 
     static handleClickDeleteCustomer = (idCustomer) => {
-      this.showLoader();
-      Services.deleteCustomer(idCustomer, user["access_token"])
-        .then((result) => {
-          setTimeout(this.hideLoader, 1000);
-          this.getAllCustomers();
-        })
-        .catch((error) => {
-          this.openToast(
-            "danger",
-            "3000",
-            error.response.data.error,
-            error.response.data.code
-          );
-        });
+      if (window.confirm("Voulez vous vraiment faire la suppression?")) {
+        this.showLoader();
+        Services.deleteCustomer(idCustomer, user["access_token"])
+          .then((result) => {
+            setTimeout(this.hideLoader, 1000);
+            this.getAllCustomers();
+            this.openToast("success","3000","Customer bien supprimé");
+          })
+          .catch((error) => {
+            this.openToast( "danger","3000",error.response.data.error,error.response.data.code);
+          });
+      }
+
     }
     static getDataCustomers() {
       return customers;
@@ -339,15 +330,21 @@ const App = () => {
         firstName: e.target.clientFirstName.value,
         email: e.target.clientEmail.value,
         address: {
+          street: e.target.clientStreet.value,
           zipCode: e.target.clientZipCode.value,
           city: e.target.clientCity.value,
         },
         mobile: e.target.clientMobile.value,
       };
-      Services.updateCustomer(id,formData,user["access_token"])
-      .then(result=>{
-        this.getAllCustomers();
-      })
+      console.log(formData);
+      Services.updateCustomer(id, formData, user["access_token"])
+        .then(result => {
+          this.getAllCustomers();
+          this.openToast("success","3000","Customer bien Modifié");
+        })
+        .catch((error)=>{
+          this.openToast( "danger","3000",error.response.data.error,error.response.data.code);
+        })
       this.handleClickCloseModalClient();
     };
 
@@ -398,7 +395,7 @@ const App = () => {
                   case "client":
                     return (
                       <>
-                        <TableClient listCustomers={listCustomers}/>
+                        <TableClient listCustomers={listCustomers} />
                       </>
                     );
                 }
@@ -433,7 +430,7 @@ const App = () => {
                   case "client":
                     return (
                       <>
-                        <TableClient listCustomers={listCustomers}/>
+                        <TableClient listCustomers={listCustomers} />
                       </>
                     );
                 }
@@ -459,7 +456,7 @@ const App = () => {
                 }
             }
           })()}
-          <FormClient show={showFormClient} showFormClientByCommande={showFormClientByCommande} isUpdateCustomer={isUpdateCustomer} formDataCustomer={selectedCustomer}/>
+          <FormClient show={showFormClient} showFormClientByCommande={showFormClientByCommande} isUpdateCustomer={isUpdateCustomer} formDataCustomer={selectedCustomer} />
           <Footer />
         </>
       )}
